got rit of some floats that caused calculation issues

+++ src/main/java/nl/dimensiontech/domotics/chargerservice/domain/ChargeSession.java
@@ -38,11 +38,11 @@ public class ChargeSession {
     private LocalDateTime startedAt;
 
     @Column(nullable = false)
-    private float startkWh;
+    private double startkWh;
 
     private LocalDateTime endedAt;
 
-    private float endkWh;
+    private double endkWh;
 
-    private float totalkwH;
+    private double totalkwH;
 }
